chore: move selectors to composites for reusability
We do not need a skeleton for these ecause the information being loaded
is not visible  and rendered for the category selector and the filtres
selector usese static data
